requirements.txt
aiocsv==1.1.1
aiofile==3.3.3
aiofiles==0.6.0
caio==0.6.3
click==7.1.2
docopt==0.6.2
Flask==1.1.2
hbmqtt==0.9.6
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paho-mqtt==1.5.1
passlib==1.7.4
PyYAML==5.4.1
queuelib==1.5.0
six==1.15.0
transitions==0.8.6
typing-extensions==3.7.4.3
websockets==8.1
Werkzeug==1.0.1